requirements.txt
binance
pandas
numpy
matplotlib
scikit-learn
python-telegram-bot==13.15
